requirements.txt
Django==3.2.16
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
requests==2.28.1
drf-yasg==1.21.4
drf-util==1.0.17
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0



